+++ src/main/java/com/github/alexisjehan/mvncheck/core/util/SystemUtils.java
@@ -23,16 +23,27 @@
  */
 package com.github.alexisjehan.mvncheck.core.util;
 
+import com.github.alexisjehan.javanilla.lang.Strings;
 import com.github.alexisjehan.javanilla.misc.quality.Ensure;
 
+import java.io.File;
 import java.nio.file.Path;
+import java.util.List;
 import java.util.Optional;
+import java.util.function.Predicate;
+import java.util.stream.Stream;
 
 /**
  * Utility class that helps to deal with the system.
  * @since 1.0.0
  */
 public final class SystemUtils {
+
+	/**
+	 * XDG directory name.
+	 * @since 2.3.0
+	 */
+	static final String XDG_DIRECTORY_NAME = "mvnchk";
 
 	/**
 	 * {@code PATH} environment variable.
@@ -45,6 +56,18 @@ public final class SystemUtils {
 	 * @since 1.0.0
 	 */
 	private static final Path USER_HOME_DIRECTORY = Path.of(System.getProperty("user.home"));
+
+	/**
+	 * Default XDG configuration user base directory.
+	 * @since 2.3.0
+	 */
+	private static final Path DEFAULT_XDG_CONFIG_USER_BASE_DIRECTORY = USER_HOME_DIRECTORY.resolve(".config");
+
+	/**
+	 * Default XDG configuration global base directory.
+	 * @since 2.3.0
+	 */
+	private static final Path DEFAULT_XDG_CONFIG_GLOBAL_BASE_DIRECTORY = Path.of("/etc/xdg");
 
 	/**
 	 * Constructor.
@@ -81,5 +104,31 @@ public final class SystemUtils {
 	 */
 	public static Path getUserHomeDirectory() {
 		return USER_HOME_DIRECTORY;
+	}
+
+	/**
+	 * Get a {@link List} of XDG configuration directories, sorted by priority.
+	 * @return the {@link List} of XDG configuration directories
+	 * @see <a href="https://specifications.freedesktop.org/basedir">https://specifications.freedesktop.org/basedir</a>
+	 * @since 2.3.0
+	 */
+	public static List<Path> getXdgConfigDirectories() {
+		return Stream.concat(
+						getEnvironmentVariable("XDG_CONFIG_HOME")
+								.filter(Predicate.not(String::isEmpty))
+								.map(Path::of)
+								.map(Stream::of)
+								.orElseGet(() -> Stream.of(DEFAULT_XDG_CONFIG_USER_BASE_DIRECTORY)),
+						getEnvironmentVariable("XDG_CONFIG_DIRS")
+								.filter(Predicate.not(String::isEmpty))
+								.map(
+										directories -> Strings.split(File.pathSeparatorChar, directories)
+												.stream()
+												.map(Path::of)
+								)
+								.orElseGet(() -> Stream.of(DEFAULT_XDG_CONFIG_GLOBAL_BASE_DIRECTORY))
+				)
+				.map(baseDirectory -> baseDirectory.resolve(XDG_DIRECTORY_NAME))
+				.toList();
 	}
 }